+++ Player/Context/Manifest/Components/Player+ManifestContext.swift
@@ -12,7 +12,7 @@ extension Player where Tech == HLSNative<ManifestContext> {
     /// Streaming extension to load and play `Manifest` sources defined by `ManifestContext`.
     ///
     /// - parameter url: the location of the media to play
-    func stream(url: URL) {
+    public func stream(url: URL) {
         let manifest = context.manifest(from: url)
         let configuration = HLSNativeConfiguration(drm: manifest.fairplayRequester)
         tech.load(source: manifest, configuration: configuration)