+++ CHANGELOG.md
@@ -13,6 +13,11 @@
 * `0.2.x` Releases - [0.2.0](#020)
 * `0.1.x` Releases - [0.1.0](#010) | [0.1.1](#011) | [0.1.2](#012) | [0.1.3](#013) | [0.1.4](#014) | [0.1.5](#015)
 
+## 2.0.90
+
+#### Changes
+* Made a couple of methods related to manifest context public.
+
 ## 2.0.89
 
 #### Bug Fixes